Project_Assignment/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
gunicorn>=21.2.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
redis>=5.0.3
email-validator>=0.0.9
